+++ sageopt/tests/test_coniclifts/test_nonlinear_operators.py
@@ -20,6 +20,8 @@ from sageopt.coniclifts.base import Variable
 from sageopt.coniclifts.operators import affine
 from sageopt.coniclifts.operators.relent import relent
 from sageopt.coniclifts.operators.norms import vector2norm
+from sageopt.coniclifts.operators.abs import abs as cl_abs
+from sageopt.coniclifts.operators.pos import pos as cl_pos
 from sageopt.coniclifts.compilers import compile_constrained_system
 from sageopt.coniclifts.cones import Cone
 
@@ -119,3 +121,43 @@ class TestNonlinearOperators(unittest.TestCase):
         assert np.all(b == b_expect)
         assert K == [Cone('+', 1), Cone('S', 4)]
 
+    def test_abs_1(self):
+        x = Variable(shape=(2,), name='x')
+        one_norm = affine.sum(cl_abs(x))
+        con = [one_norm <= 5]
+        A_expect = np.array([[0,  0, -1, -1],
+                             [1,  0, 1,  0],
+                             [-1, 0, 1,  0],
+                             [0,  1, 0,  1],
+                             [0, -1, 0,  1]])
+        A, b, K, _1, _2, _3 = compile_constrained_system(con)
+        A = np.round(A.toarray(), decimals=1)
+        assert np.all(A == A_expect)
+        assert np.all(b == np.array([5, 0, 0, 0, 0]))
+        assert K == [Cone('+', 1), Cone('+', 2), Cone('+', 2)]
+        # value propagation
+        x.value = np.array([1, -2])
+        viol = con[0].violation()
+        assert viol == 0
+        x.value = np.array([-3, 3])
+        viol = con[0].violation()
+        assert viol == 1
+
+    def test_pos_1(self):
+        x = Variable(shape=(3,), name='x')
+        con = [cl_pos(affine.sum(x)-7) <= 5]
+        A_expect = np.array([[0,   0,  0, -1],
+                             [0,   0,  0,  1],
+                             [-1, -1, -1,  1]])
+        A, b, K, _1, _2, _3 = compile_constrained_system(con)
+        A = np.round(A.toarray(), decimals=1)
+        assert np.all(A == A_expect)
+        assert np.all(b == np.array([5, 0, 7]))
+        assert K == [Cone('+', 1), Cone('+', 2)]
+        # value propagation
+        x.value = np.array([4, 4, 4])
+        viol = con[0].violation()
+        assert viol == 0
+        x.value = np.array([4.2, 3.9, 4.0])
+        viol = con[0].violation()
+        assert abs(viol - 0.1) < 1e-7